refactor(rooch-da): restructure OpenDA backend architecture
Reorganized OpenDA backend to enhance modularity and maintainability. Split `new_operator` into its own function, restructured module imports, and relocated `backend.rs` to `openda_backend.rs`. Improved error handling in segment submission logic.
