+++ crates/rooch-da/src/backend/openda/openda_backend.rs
@@ -4,7 +4,7 @@
 use crate::backend::openda::avail::AvailFusionClientConfig;
 use crate::backend::openda::celestia::{CelestiaClient, WrappedNamespace};
 use crate::backend::openda::opendal::BACK_OFF_MIN_DELAY;
-use crate::backend::openda::operator::{Operator, OperatorConfig};
+use crate::backend::openda::operator::{new_operator, Operator, OperatorConfig};
 use crate::backend::DABackend;
 use async_trait::async_trait;
 use opendal::layers::{LoggingLayer, RetryLayer};
@@ -14,13 +14,6 @@ use rooch_types::da::batch::DABatch;
 use rooch_types::da::chunk::{Chunk, ChunkV0};
 use std::collections::HashMap;
 use std::sync::Arc;
-
-#[async_trait]
-impl DABackend for OpenDABackend {
-    async fn submit_batch(&self, batch: Arc<DABatch>) -> anyhow::Result<()> {
-        self.pub_batch(batch).await
-    }
-}
 
 pub struct OpenDABackend {
     operator_config: OperatorConfig,
@@ -41,8 +34,11 @@ impl OpenDABackend {
             operator,
         })
     }
+}
 
-    pub async fn pub_batch(&self, batch: Arc<DABatch>) -> anyhow::Result<()> {
+#[async_trait]
+impl DABackend for OpenDABackend {
+    async fn submit_batch(&self, batch: Arc<DABatch>) -> anyhow::Result<()> {
         let chunk: ChunkV0 = (*batch).clone().into();
 
         let scheme = self.operator_config.scheme.clone();
@@ -79,46 +75,4 @@ impl OpenDABackend {
 
         Ok(())
     }
-}
-
-async fn new_operator(
-    operator_config: OperatorConfig,
-    scheme_config: HashMap<String, String>,
-) -> anyhow::Result<Box<dyn Operator>> {
-    let max_retries = operator_config.max_retries;
-    let scheme = operator_config.scheme.clone();
-
-    let operator: Box<dyn Operator> = match scheme {
-        OpenDAScheme::Avail => {
-            let avail_fusion_config =
-                AvailFusionClientConfig::from_scheme_config(scheme_config, max_retries)?;
-            let avail_fusion_client = avail_fusion_config.build_client()?;
-            Box::new(avail_fusion_client)
-        }
-        OpenDAScheme::Celestia => {
-            let namespace = WrappedNamespace::from_string(&operator_config.namespace.clone())?;
-            Box::new(
-                CelestiaClient::new(
-                    namespace.into_inner(),
-                    &scheme_config["endpoint"],
-                    scheme_config.get("auth_token").map(|s| s.as_str()),
-                    max_retries,
-                )
-                .await?,
-            )
-        }
-        _ => {
-            let mut op = opendal::Operator::via_iter(Scheme::from(scheme), scheme_config)?;
-            op = op
-                .layer(
-                    RetryLayer::new()
-                        .with_max_times(max_retries)
-                        .with_min_delay(BACK_OFF_MIN_DELAY),
-                )
-                .layer(LoggingLayer::default());
-            op.check().await?;
-            Box::new(op)
-        }
-    };
-    Ok(operator)
 }
